+++ src/format_request.js
@@ -10,6 +10,12 @@ import getFieldAttributes from './utils/field_attributes.js';
 import extend from './utils/extend.js';
 import buildQuery from './get.js';
 
+/* eslint-disable jsdoc/valid-types */
+/**
+ * @typedef {import('./index.js').default} Dare
+ */
+/* eslint-enable jsdoc/valid-types */
+
 /**
  * Format Request initiation
  *
@@ -21,17 +27,11 @@ export default function (options) {
 }
 
 /**
- * @typedef {object} Dare
- * @param {object} options - instance options
- * @param {Function} table_alias_handler - The db table this references
- */
-
-/**
  * Format Request
  *
  * @param {object} options - Current iteration
  * @param {Dare} dareInstance - Instance of Dare
- * @returns {object} formatted object with all the joins
+ * @returns {Promise<object>} formatted object with all the joins
  */
 async function format_request(options, dareInstance) {
 	if (!options) {
@@ -110,7 +110,7 @@ async function format_request(options, dareInstance) {
 			derivedMethod in model
 				? model[derivedMethod]
 				: // Or use the default model
-				  models?.default?.[derivedMethod];
+					models?.default?.[derivedMethod];
 
 		if (handler) {
 			// Trigger the handler which alters the options...
@@ -126,14 +126,18 @@ async function format_request(options, dareInstance) {
 	 */
 	{
 		Object.keys(table_schema).forEach(key => {
-			const {defaultValue = {}} = getFieldAttributes(key, table_schema);
+			const {defaultValue} = getFieldAttributes(
+				key,
+				table_schema,
+				dareInstance
+			);
 
 			/*
 			 * Check the defaultValue for the method has been assigned
 			 * -> That there is no definition for the value in the filter and join options
 			 * -> That we're trying to get their original field names
 			 */
-			if (method in defaultValue) {
+			if (defaultValue !== undefined) {
 				// Does the fields exist?
 				const filterHasKey = Object.keys({
 					...options.filter,
@@ -149,7 +153,7 @@ async function format_request(options, dareInstance) {
 				// If there is no match
 				if (!filterHasKey) {
 					// Extend the join object with the default value
-					extend(options, {join: {[key]: defaultValue[method]}});
+					extend(options, {join: {[key]: defaultValue}});
 				}
 			}
 		});
@@ -302,7 +306,7 @@ async function format_request(options, dareInstance) {
 		const extract = extractJoined.bind(null, 'groupby', true);
 
 		// Set reducer options
-		const reducer = groupbyReducer({current_path, extract, table_schema});
+		const reducer = groupbyReducer({current_path, extract});
 
 		// Return array of immediate props
 		options.groupby = toArray(options.groupby).reduce(reducer, []);
@@ -512,7 +516,7 @@ async function format_request(options, dareInstance) {
 			const sub_query = buildQuery(options, dareInstance);
 
 			sql_where_conditions = [
-				SQL`${raw(parentReferences)}
+				SQL`${raw(parentReferences[0])}
 				NOT IN (
 					SELECT ${raw(options.fields)} FROM (
 						${sub_query}