+++ src/index.js
@@ -16,6 +16,58 @@ import format_request from './format_request.js';
 
 import response_handler, {responseRowHandler} from './response_handler.js';
 
+/* eslint-disable jsdoc/valid-types */
+/**
+ * @typedef {import('sql-template-tag').Sql} Sql
+ *
+ * @typedef {object} Model
+ * @property {Object<string, object | Function | Array<string> | string | null | boolean>} [schema] - Model Schema
+ * @property {string} [table] - Alias for the table
+ * @property {Object<string, string>} [shortcut_map] - Shortcut map
+ * @property {Function} [get] - Get handler
+ * @property {Function} [post] - Post handler
+ * @property {Function} [patch] - Patch handler
+ * @property {Function} [del] - Delete handler
+ *
+ * @typedef {object} RequestObject
+ * @property {string} [table] - Name of the table to query
+ * @property {Array} [fields] - Fields array to return
+ * @property {object} [filter] - Filter Object to query
+ * @property {object} [join] - Place filters on the joining tables
+ * @property {object} [body] - Body containing new data
+ * @property {RequestObject} [query] - Query attached to a post request to create INSERT...SELECT operations
+ * @property {number} [limit] - Number of items to return
+ * @property {number} [start] - Number of items to skip
+ * @property {string | string[]} [orderby] - Array of fields to order by
+ * @property {string | string[]} [groupby] - Field to group by
+ * @property {string} [duplicate_keys] - 'ignore' to prevent throwing Duplicate key errors
+ * @property {string[]} [duplicate_keys_update] - An array of fields to update on presence of duplicate key constraints
+ * @property {*} [notfound] - If not undefined will be returned in case of a single entry not found
+ * @property {Object<string, Model>} [models] - Models with schema defintitions
+ * @property {Function} [validateInput] - Validate input
+ * @property {boolean} [infer_intermediate_models] - Infer intermediate models
+ * @property {Function} [rowHandler] - Override default Function to handle each row
+ * @property {Function} [getFieldKey] - Override default Function to interpret the field key
+ * @property {string} [conditional_operators_in_value] - Allowable conditional operators in value
+ * @property {any} [state] - Arbitary data to carry through to the model/response handlers
+ *
+ * @typedef {object} InternalProps
+ * @property {'post' | 'get' | 'patch' | 'del'} [method] - Method to use
+ * @property {string} [name] - Model Name derived
+ * @property {boolean} [single] - Return a single item
+ * @property {boolean} [skip] - Skip the request
+ * @property {string} [alias] - Alias for the table
+ * @property {boolean} [countRows] - Count all rows
+ * @property {string} [sql_table] - SQL Table
+ * @property {string} [sql_alias] - SQL Alias
+ * @property {Array} [sql_joins] - SQL Join
+ * @property {string} [ignore] - SQL Fields
+ * @property {Array} [sql_where_conditions] - SQL Where conditions
+ *
+ * @typedef {RequestObject & InternalProps} QueryOptions
+ */
+/* eslint-enable jsdoc/valid-types */
+
 /*
  * Export Dare Error object
  */
@@ -25,8 +77,8 @@ export {DareError};
  * Dare
  * Sets up a new instance of Dare
  *
- * @param {object} options - Initial options defining the instance
- * @returns {object} instance of dare
+ * @param {QueryOptions} options - Initial options defining the instance
+ * @returns {Dare} instance of dare
  */
 function Dare(options = {}) {
 	// Overwrite default properties
@@ -41,7 +93,13 @@ export default Dare;
 Dare.DareError = DareError;
 
 // Set default function
-Dare.prototype.execute = async () => {
+/**
+ * Set default execution handler
+ * @param {object} requestQuery - Request object
+ * @returns {Promise<object>} Response
+ */
+// eslint-disable-next-line no-unused-vars
+Dare.prototype.execute = async requestQuery => {
 	throw new DareError(
 		DareError.INVALID_SETUP,
 		'Define dare.execute to continue'
@@ -49,15 +107,22 @@ Dare.prototype.execute = async () => {
 };
 
 // Group concat
+/** @type {string} */
 Dare.prototype.group_concat = '$$';
 
 // Rowid, name of primary key field used in grouping operations
+/** @type {string} */
 Dare.prototype.rowid = '_rowid';
 
 // Set the Max Limit for SELECT statements
+/** @type {number} */
 Dare.prototype.MAX_LIMIT = null;
 
+// Capture the generated field functions to run after the request
+Dare.prototype.generated_fields = [];
+
 // Default options
+/** @type {Partial<QueryOptions>} */
 Dare.prototype.options = {
 	// Infer intermediate tables when two models are not directly linked
 	infer_intermediate_models: true,
@@ -83,6 +148,8 @@ Dare.prototype.get_unique_alias = function (iterate = 1) {
 	return `\`${str}\``;
 };
 
+// eslint-disable-next-line jsdoc/valid-types
+/** @type {(options: QueryOptions) => Promise<QueryOptions>} */
 Dare.prototype.format_request = format_request;
 
 Dare.prototype.response_handler = response_handler;
@@ -91,7 +158,7 @@ Dare.prototype.response_handler = response_handler;
  * GetFieldKey
  * @param {string} field - Field
  * @param {object} schema - Model Schema
- * @returns {object | undefined} Field Key
+ * @returns {string | void} Field Key
  */
 // eslint-disable-next-line no-unused-vars
 Dare.prototype.getFieldKey = function getFieldKey(field, schema) {
@@ -99,12 +166,70 @@ Dare.prototype.getFieldKey = function getFieldKey(field, schema) {
 };
 
 /**
+ * Fulltext Parser
+ * This will format a string to make it compliant with MySQL Fulltext search
+ * Such as wrapping special characters in quotes where they appear in the middle of words
+ * Removing any trailing '*' characters which succeed a quoted string
+ * e.g. `+test@example.com*` becomes `+"test@example.com"`
+ * @param {string} input - Input string
+ * @returns {string} Formatted string
+ */
+Dare.prototype.fulltextParser = function fulltextParser(input) {
+	function safequote(text) {
+		let suffix = '';
+
+		if (text.endsWith('*')) {
+			suffix = '*';
+			text = text.slice(0, -1);
+		}
+
+		if (text.match(/['@-]/)) {
+			return `"${text}"`;
+		}
+		return text + suffix;
+	}
+
+	if (typeof input !== 'string' || input === '') {
+		throw new DareError(
+			DareError.INVALID_REQUEST,
+			'Fulltext input must be a string'
+		);
+	}
+
+	// Replace any special characters with quotes
+	const resp = input.matchAll(
+		/\s*(?<sign>[+<>~-]?)(?:\((?<subexpression>.*?)\)|(?<quoted>".*?")|(?<unquoted>\S+))(?<suffix>\*?)/g
+	);
+	const output = [...resp]
+		.filter(({groups: {subexpression, quoted, unquoted}}) =>
+			quoted
+				? quoted.length > 2
+				: subexpression || unquoted.replace(/^[*+-]+/, '')
+		)
+		.map(({groups: {sign, subexpression, quoted, unquoted, suffix}}) => {
+			if (subexpression) {
+				return `${sign}(${this.fulltextParser(subexpression)})`;
+			} else if (quoted) {
+				return `${sign}${quoted}`;
+			} else {
+				return `${sign}${safequote(unquoted + suffix)}`;
+			}
+		});
+
+	return output.join(' ');
+};
+
+/* eslint-disable jsdoc/valid-types */
+/* eslint-disable jsdoc/check-tag-names */
+/**
  * Dare.after
  * Defines where the instance goes looking to apply post execution handlers and potentially mutate the response
- *
- * @param {object|Array} resp - Response object
- * @returns {object} response data formatted or not
+ * @template {object|Array} T
+ * @param {T} resp - Response object
+ * @returns {T} response data formatted or not
  */
+/* eslint-enable jsdoc/valid-types */
+/* eslint-enable jsdoc/check-tag-names */
 Dare.prototype.after = function (resp) {
 	// Define the after handler
 	const handler = `after${this.options.method.replace(/^[a-z]/, m =>
@@ -124,13 +249,16 @@ Dare.prototype.after = function (resp) {
 /**
  * Use
  * Creates a new instance of Dare and merges new options with the base options
- * @param {object} options - set of instance options
- * @returns {object} Instance of Dare
+ * @param {QueryOptions} options - set of instance options
+ * @returns {Dare} Instance of Dare
  */
 Dare.prototype.use = function (options = {}) {
 	const inst = Object.create(this);
 
-	// Create a new options, merging inheritted and new
+	/**
+	 * Create a new options, merging inheritted and new
+	 * @type {QueryOptions} inst.options
+	 */
 	inst.options = extend(clone(this.options), options);
 
 	// Define the Row handler to format the results
@@ -150,8 +278,15 @@ Dare.prototype.use = function (options = {}) {
 };
 
 /**
+ * Define a resultset
+ * @type {Array<object>} resultset
+ */
+Dare.prototype.resultset = undefined;
+
+/**
  * Add a row to the resultset
  * @param {object} row - Row record to add to the rows resultset
+ * @returns {void}
  */
 Dare.prototype.addRow = function (row) {
 	// Format the SQL Row
@@ -168,15 +303,17 @@ Dare.prototype.addRow = function (row) {
  * Dare.sql
  * Prepares and processes SQL statements
  *
- * @param {string} sql - SQL string containing the query
- * @param {Array<Array, string, number, boolean>} values - List of prepared statement values
- * @returns {Promise<object|Array>} Returns response object or array of values
+ * @param {string | Sql | {sql: string, values: Array}} sql - SQL string containing the query
+ * @param {Array} [values] - List of prepared statement values
+ * @returns {Promise} Returns response object or array of values
  */
 Dare.prototype.sql = async function sql(sql, values) {
-	let req = {sql, values};
+	let req;
 
 	if (typeof sql === 'object') {
 		req = sql;
+	} else {
+		req = {sql, values};
 	}
 
 	const resp = await this.execute(req);
@@ -187,24 +324,27 @@ Dare.prototype.sql = async function sql(sql, values) {
  * Dare.get
  * Triggers a DB SELECT request to rerieve records from the database.
  *
- * @param {string} table - Name of the table to query
- * @param {Array} fields - Fields array to return
- * @param {object} filter - Filter Object to query
- * @param {object} opts - An Options object containing all other request options
- * @returns {Promise<object|Array>} Results
+ * @param {string | RequestObject} table - Name of the table to query
+ * @param {Array} [fields] - Fields array to return
+ * @param {object} [filter] - Filter Object to query
+ * @param {RequestObject} [options] - An Options object containing all other request options
+ * @returns {Promise<any>} Results
  */
-Dare.prototype.get = async function get(table, fields, filter, opts = {}) {
-	// Get Request Object
-	if (typeof table === 'object') {
-		opts = {...table};
-	} else {
-		// Shuffle
-		if (typeof fields === 'object' && !Array.isArray(fields)) {
-			// Fields must be defined
-			throw new DareError(DareError.INVALID_REQUEST);
-		}
+Dare.prototype.get = async function get(table, fields, filter, options = {}) {
+	/**
+	 * @type {QueryOptions} opts
+	 */
+	const opts =
+		typeof table === 'object'
+			? // Clone
+				{...table}
+			: // Clone and extend
+				{...options, table, filter, fields};
 
-		opts = {...opts, table, fields, filter};
+	// Ensure fields is provided
+	if (typeof fields === 'object' && !Array.isArray(fields)) {
+		// Fields must be defined
+		throw new DareError(DareError.INVALID_REQUEST);
 	}
 
 	// Define method
@@ -247,20 +387,21 @@ Dare.prototype.get = async function get(table, fields, filter, opts = {}) {
  * Dare.getCount
  * Returns the total number of results which match the conditions
  *
- * @param {string} table - Name of the table to query
- * @param {object} filter - Filter Object to query
- * @param {object} opts - An Options object containing all other request options
- * @returns {Promise<integer>} Number of matched items
+ * @param {string | RequestObject} table - Name of the table to query
+ * @param {object} [filter] - Filter Object to query
+ * @param {RequestObject} [options] - An Options object containing all other request options
+ * @returns {Promise<number>} Number of matched items
  */
-Dare.prototype.getCount = async function getCount(table, filter, opts = {}) {
-	// Get Request Object
-	if (typeof table === 'object') {
-		// Clone
-		opts = {...table};
-	} else {
-		// Clone and extend
-		opts = {...opts, table, filter};
-	}
+Dare.prototype.getCount = async function getCount(table, filter, options = {}) {
+	/**
+	 * @type {QueryOptions} opts
+	 */
+	const opts =
+		typeof table === 'object'
+			? // Clone
+				{...table}
+			: // Clone and extend
+				{...options, table, filter};
 
 	// Define method
 	opts.method = 'get';
@@ -295,20 +436,22 @@ Dare.prototype.getCount = async function getCount(table, filter, opts = {}) {
  * Dare.patch
  * Updates records matching the conditions
  *
- * @param {string} table - Name of the table to query
- * @param {object} filter - Filter Object to query
- * @param {object} body - Body containing new data
- * @param {object} [opts] - An Options object containing all other request options
- * @param {string} [opts.duplicate_keys] - 'ignore' to prevent throwing Duplicate key errors
- * @param {number} [opts.limit=1] - Number of items to change
- * @returns {Promise<object>} Affected Rows statement
+ * @param {string | RequestObject} table - Name of the table to query
+ * @param {object} [filter] - Filter Object to query
+ * @param {object} [body] - Body containing new data
+ * @param {RequestObject} [options] - An Options object containing all other request options
+ * @returns {Promise<any>} Affected Rows statement
  */
-Dare.prototype.patch = async function patch(table, filter, body, opts = {}) {
-	// Get Request Object
-	opts =
+Dare.prototype.patch = async function patch(table, filter, body, options = {}) {
+	/**
+	 * @type {QueryOptions} opts
+	 */
+	const opts =
 		typeof table === 'object'
-			? table
-			: Object.assign(opts, {table, filter, body});
+			? // Clone
+				{...table}
+			: // Clone and extend
+				{...options, table, filter, body};
 
 	// Define method
 	opts.method = 'patch';
@@ -374,18 +517,21 @@ Dare.prototype.patch = async function patch(table, filter, body, opts = {}) {
  * Dare.post
  * Insert new data into database
  *
- * @param {string} table - Name of the table to query
- * @param {object|Array<objects>} body - Body containing new data
- * @param {object} [opts] - An Options object containing all other request options
- * @param {Array} [opts.duplicate_keys_update] - An array of fields to update on presence of duplicate key constraints
- * @param {string} [opts.duplicate_keys] - 'ignore' to prevent throwing Duplicate key errors
- * @returns {Promise<object>} Affected Rows statement
+ * @param {string | RequestObject} table - Name of the table to query
+ * @param {object | Array<object>} [body] - Body containing new data
+ * @param {RequestObject} [options] - An Options object containing all other request options
+ * @returns {Promise<any>} Affected Rows statement
  */
-
-Dare.prototype.post = async function post(table, body, opts = {}) {
-	// Get Request Object
-	opts =
-		typeof table === 'object' ? table : Object.assign(opts, {table, body});
+Dare.prototype.post = async function post(table, body, options = {}) {
+	/**
+	 * @type {QueryOptions} opts
+	 */
+	const opts =
+		typeof table === 'object'
+			? // Clone
+				{...table}
+			: // Clone and extend
+				{...options, table, body};
 
 	// Post
 	opts.method = 'post';
@@ -531,20 +677,11 @@ Dare.prototype.post = async function post(table, body, opts = {}) {
 						return;
 					}
 
-					/*
-					 * Get the default Value of the post operation
-					 * -> Reassign it back to the fieldAttributes.defaultValue, no need for the others
-					 */
-					const defaultValue = fieldAttributes.defaultValue?.post;
-					if (fieldAttributes.defaultValue) {
-						Object.assign(fieldAttributes, {defaultValue});
-					}
-
 					// Validate with an undefined value
 					validateInput?.(fieldAttributes, field);
 
 					// Default values?
-					if (defaultValue) {
+					if (fieldAttributes.defaultValue) {
 						// Get the index in the field list
 						let i = fields.indexOf(field);
 
@@ -555,7 +692,7 @@ Dare.prototype.post = async function post(table, body, opts = {}) {
 						}
 
 						// Insert the defaultValue at that position
-						_data[i] = defaultValue;
+						_data[i] = fieldAttributes.defaultValue;
 					}
 				}
 			});
@@ -613,17 +750,21 @@ Dare.prototype.post = async function post(table, body, opts = {}) {
  * Dare.del
  * Delete a record matching condition
  *
- * @param {string} table - Name of the table to query
- * @param {object} filter - Filter Object to query
- * @param {object} opts - An Options object containing all other request options
- * @returns {Promise<object>} Affected Rows statement
+ * @param {string | RequestObject} table - Name of the table to query
+ * @param {object} [filter] - Filter Object to query
+ * @param {RequestObject} [options] - An Options object containing all other request options
+ * @returns {Promise<any>} Affected Rows statement
  */
-Dare.prototype.del = async function del(table, filter, opts = {}) {
-	// Get Request Object
-	opts =
+Dare.prototype.del = async function del(table, filter, options = {}) {
+	/**
+	 * @type {QueryOptions} opts
+	 */
+	const opts =
 		typeof table === 'object'
-			? table
-			: Object.assign(opts, {table, filter});
+			? // Clone
+				{...table}
+			: // Clone and extend
+				{...options, table, filter};
 
 	// Delete
 	opts.method = 'del';
@@ -722,7 +863,7 @@ function onDuplicateKeysUpdate(keys = []) {
  * @param {Function} [obj.validateInput] - Custom validation function
  * @param {object} obj.dareInstance - Dare Instance
  * @throws Will throw an error if the field is not writable
- * @returns {string} A singular value which can be inserted
+ * @returns {{field: string, value: *}} A singular value which can be inserted
  */
 function formatInputValue({
 	tableSchema = {},
@@ -777,7 +918,7 @@ function formatInputValue({
 	}
 
 	// Check this is not an object
-	if (value && typeof value === 'object') {
+	if (value && typeof value === 'object' && !Buffer.isBuffer(value)) {
 		throw new DareError(
 			DareError.INVALID_VALUE,
 			`Field '${field}' does not accept objects as values: '${JSON.stringify(
@@ -803,12 +944,12 @@ function formatInputValue({
 /**
  * Return un-aliased field names
  *
- * @param {object} [tableSchema={}] - An object containing the table schema
+ * @param {object} tableSchema - An object containing the table schema
  * @param {string} field - field identifier
  * @param {object} dareInstance - Dare Instance
  * @returns {string} Unaliased field name
  */
-function unAliasFields(tableSchema = {}, field, dareInstance) {
+function unAliasFields(tableSchema, field, dareInstance) {
 	const {alias} = getFieldAttributes(field, tableSchema, dareInstance);
 	return alias || field;
 }
@@ -816,8 +957,8 @@ function unAliasFields(tableSchema = {}, field, dareInstance) {
 /**
  * SetDefaultNotFoundHandler
  * As the name suggests
- * @param {object} opts - request options
- * @returns {void}
+ * @param {QueryOptions} opts - query options
+ * @returns {QueryOptions} query options
  */
 function setDefaultNotFoundHandler(opts) {
 	if (!('notfound' in opts)) {