+++ test/specs/init.spec.js
@@ -1,3 +1,4 @@
+import {expect} from 'chai';
 import Dare, {DareError} from '../../src/index.js';
 import clone from 'tricks/object/clone.js';
 
@@ -8,7 +9,9 @@ describe('Dare', () => {
 	});
 
 	it('should define default options', () => {
-		const models = {};
+		const models = {
+			mytable: {},
+		};
 		const dare = new Dare({
 			models,
 		});
@@ -22,7 +25,7 @@ describe('Dare', () => {
 	it('should throw errors if dare.execute is not defined', () => {
 		const dare = new Dare();
 
-		const test = dare.sql({sql: 'SELECT 1=1'});
+		const test = dare.sql('SELECT 1=1');
 
 		return expect(test).to.be.eventually.rejectedWith(
 			DareError,